refactor(stm): remove the in-process IVC prover setup cache
