+++ mithril-stm/src/proof_system/ivc_halo2_snark/mod.rs
@@ -1,10 +1,8 @@
 pub(crate) mod errors;
-pub(crate) mod ivc_prover_setup_cache;
 pub(crate) mod proof;
 mod prover_input;
 mod prover_setup;
 pub(crate) mod rolling_state;
 pub(crate) mod verifier_setup;
 
-pub(crate) use ivc_prover_setup_cache::load_ivc_prover_setup;
 pub(crate) use prover_setup::IvcSnarkProverSetup;
